move tests to tests/e2e; move everything into same folder (reports etc.)
