+++ .gitignore
@@ -17,7 +17,7 @@ dist-firefox-v2
 public/manifest.json
 *.local
 coverage
-/tests/.env
+.env
 .husky
 .vscode
 
@@ -33,8 +33,7 @@ coverage
 *.sw?
 
 # playwright
-/test-results/
-/test-results/.auth
-/playwright-report/
-/blob-report/
+/tests/e2e/test-results/
+/tests/e2e/test-results/.auth
+/tests/e2e/playwright-report/
 /playwright/.cache/
