+++ tests/e2e/helpers.ts
@@ -1,8 +1,8 @@
 import path from 'node:path';
 import { readFileSync } from 'node:fs';
-import { SRC_DIR } from '../esbuild/config';
+import { SRC_DIR } from '../../esbuild/config';
 
-type TranslationKeys = keyof typeof import('../src/_locales/en/messages.json');
+type TranslationKeys = keyof typeof import('../../src/_locales/en/messages.json');
 
 type TranslationData = Record<
   TranslationKeys,
