+++ tsconfig.json
@@ -29,7 +29,7 @@
     "./jest.config.ts",
     "./jest.setup.ts",
     "./scripts/**/*",
-    "./tests/**/*",
+    "./tests/e2e/**/*/*",
     "./.github/actions/*.cjs",
     "esbuild/**/*"
   ],
